+++ src/components/FormElement/FormElement.styles.ts
@@ -49,5 +49,20 @@ export const getStyles = (theme: GrafanaTheme2) => {
       display: flex;
       align-items: center;
     `,
+    checkboxWrap: css`
+      width: 100%;
+      & > div {
+        flex: auto;
+      }
+    `,
+    checkboxRow: css`
+      display: flex;
+      flex-wrap: wrap;
+      flex-direction: column;
+      align-items: flex-start;
+    `,
+    checkbox: css`
+      padding: ${theme.spacing(1)};
+    `,
   };
 };