+++ src/components/FormElement/FormElement.tsx
@@ -1,6 +1,7 @@
 import { cx } from '@emotion/css';
 import { DateTime, dateTime } from '@grafana/data';
 import {
+  Checkbox,
   DatePickerWithInput,
   DateTimePicker,
   FileDropzone,
@@ -398,7 +399,7 @@ export const FormElement: React.FC<Props> = ({ element, onChange, highlightClass
           <FileDropzone
             options={{
               accept: element.accept || undefined,
-              multiple: true,
+              multiple: element.multiple,
               onDrop: (files: File[]) => {
                 onChange<typeof element>({
                   ...element,
@@ -433,6 +434,38 @@ export const FormElement: React.FC<Props> = ({ element, onChange, highlightClass
             >
               {element.linkText || element.value}
             </TextLink>
+          </div>
+        </InlineField>
+      )}
+
+      {element.type === FormElementType.CHECKBOX_LIST && (
+        <InlineField
+          label={element.title}
+          grow={!element.width}
+          labelWidth={applyWidth(element.labelWidth)}
+          tooltip={element.tooltip}
+          transparent={!element.title}
+          disabled={element.disabled}
+          className={styles.checkboxWrap}
+          data-testid={TEST_IDS.formElements.fieldCheckboxListContainer}
+        >
+          <div style={{ width: element.width ? theme.spacing(element.width) : 'auto' }} className={styles.checkboxRow}>
+            {element.options.map((option) => (
+              <Checkbox
+                className={styles.checkbox}
+                key={option.id}
+                value={element.value.some((val) => val === option.value)}
+                label={option.label}
+                onChange={(event) => {
+                  onChange<typeof element>({
+                    ...element,
+                    value: event.currentTarget.checked
+                      ? element.value.concat(option.value)
+                      : element.value.filter((value) => value !== option.value),
+                  });
+                }}
+              />
+            ))}
           </div>
         </InlineField>
       )}